Corrections made to joystick class + implemented destructors

+++ include/joystick.h
@@ -5,6 +5,7 @@
 #ifndef PROJECT_CTRL_GAMING_CONTROLLER_JOYSTICK_H
 #define PROJECT_CTRL_GAMING_CONTROLLER_JOYSTICK_H
 #include <stdint.h>
+#include "../ArduinoXInput/src/XInput.h"
 
 class Joystick {
 private:
@@ -14,9 +15,9 @@ private:
     uint8_t analogYPin;
 public:
     Joystick(uint8_t analogXPin, uint8_t analogYPin);
-    void move(double x, double y);
-    double getX() const;
-    double getY() const;
+    void move(int16_t x, int16_t y);
+    int16_t getX() const;
+    int16_t getY() const;
     int32_t readAnalogX();
     int32_t readAnalogY();
     void setX(double newX);
